divide db into multiple packages and bump to 1.18
