+++ pkg/data/db/db.go
@@ -17,120 +17,51 @@ package db
 // A ORM library is used to connect to the database: see https://gorm.io/docs/
 
 import (
-	"errors"
 	"fmt"
-	"os"
 
-	"gorm.io/driver/postgres"
-	"gorm.io/driver/sqlite"
-	"gorm.io/gorm"
+	"carbonaut.cloud/carbonaut/pkg/data/db/methods"
+	"carbonaut.cloud/carbonaut/pkg/data/db/provider/postgres"
+	"carbonaut.cloud/carbonaut/pkg/data/db/provider/sqlite"
 )
 
-type PostgresConfig struct {
-	Port         int
-	Password     string
-	Host         string
-	User         string
-	DatabaseName string
-	SSLMode      SSLMode
+func ValidateConfig(cfg *Config) error {
+	return fmt.Errorf("provided database configuration is invalid")
 }
 
-type SQLiteConfig struct {
-	DatabaseFileName string
+type Config struct {
+	Provider string `validate:"nonzero"`
+	// Provider config gets dynamically set over the provider string specification
+	ProviderConfig interface{}
 }
 
-type SSLMode string
+// establish a connection to the configured database
+func Connect(cfg *Config) (methods.ICarbonDB, error) {
+	provider, err := resolveProvider(cfg)
+	if err != nil {
+		return nil, err
+	}
+	provider.Connect(cfg)
+	return nil, fmt.Errorf("not implemented")
+}
 
-const (
-	SSLModeDisable = "disable"
-	SSLModeEnable  = "enable"
-)
+// resolve which database provider is specified in the configuration
+func resolveProvider(cfg *Config) (IProvider, error) {
+	switch cfg.Provider {
+	case sqlite.Provider.Name:
+		fmt.Println("OS X.")
+	case postgres.Provider.Name:
+		fmt.Println("OS X.")
+		// postgres.Provider.Connect(cfg.ProviderConfig)
+	default:
+		return nil, fmt.Errorf("specified provider %s is not supported", cfg.Provider)
+	}
+	return nil, fmt.Errorf("not implemented")
+}
 
 // DatabaseDriver
 type DatabaseDriver string
 
-func ConnectToSQLite(cfg *SQLiteConfig) (ICarbonDB, error) {
-	if _, err := os.Stat(cfg.DatabaseFileName); errors.Is(err, os.ErrNotExist) {
-		return nil, err
-	}
-	// open connection to db file
-	db, err := gorm.Open(sqlite.Open(cfg.DatabaseFileName), &gorm.Config{})
-	if err != nil {
-		return nil, err
-	}
-	// migrate tables
-	carbonDB := carbonDB{db}
-	if err := carbonDB.Migrate(); err != nil {
-		return nil, err
-	}
-	return carbonDB, nil
-}
-
-// Connect establishes a connection to the provided database configuration
-// The database can hosted locally (or somewhere else)
-// Locally you can use the officially postgres container image
-// 1. podman run -it --rm -p 127.0.0.1:5432:5432/tcp -e POSTGRES_PASSWORD=test postgres
-// 2. psql -d postgres -h localhost -U postgres
-// 3. enter password: test
-// Setting the same information in PostgresConfig to connect to the local hosted database
-func ConnectToPostgres(cfg *PostgresConfig) (ICarbonDB, error) {
-	// open connection to db
-	db, err := gorm.Open(postgres.Open(
-		fmt.Sprintf(
-			"host=%s user=%s password=%s dbname=%s port=%d sslmode=%s",
-			cfg.Host, cfg.User, cfg.Password, cfg.DatabaseName, cfg.Port, cfg.SSLMode,
-		)), &gorm.Config{})
-	if err != nil {
-		return nil, err
-	}
-	// migrate tables
-	carbonDB := carbonDB{db}
-	if err := carbonDB.Migrate(); err != nil {
-		return nil, err
-	}
-	return carbonDB, nil
-}
-
-type ICarbonDB interface {
-	Get(id uint) (*Emissions, error)
-	Save(emissions *Emissions) error
-	List(offset, limit int) ([]*Emissions, error)
-	Delete(id uint) error
-	Migrate() error
-	SearchByResourceName(q string, offset, limit int) ([]*Emissions, error)
-}
-
-type carbonDB struct {
-	db *gorm.DB
-}
-
-func (d carbonDB) Get(id uint) (*Emissions, error) {
-	e := &Emissions{}
-	err := d.db.Where(`id = ?`, id).Find(e).Error
-	return e, err
-}
-
-func (d carbonDB) Save(emissions *Emissions) error {
-	return d.db.Save(emissions).Error
-}
-
-func (d carbonDB) List(offset, limit int) ([]*Emissions, error) {
-	var l []*Emissions
-	err := d.db.Offset(offset).Limit(limit).Find(&l).Error
-	return l, err
-}
-
-func (d carbonDB) Delete(id uint) error {
-	return d.db.Delete(&Emissions{ID: id}).Error
-}
-
-func (d carbonDB) Migrate() error {
-	return d.db.AutoMigrate(&Emissions{})
-}
-
-func (d carbonDB) SearchByResourceName(q string, offset, limit int) ([]*Emissions, error) {
-	var l []*Emissions
-	err := d.db.Where(`resource_name like ?`, "%"+q+"%").
-		Offset(offset).Limit(limit).Find(&l).Error
-	return l, err
+type IProvider interface {
+	Connect(cfg interface{}) (methods.ICarbonDB, error)
+	Validate(cfg interface{}) error
 }
