+++ pkg/data/db/provider/postgres/postgres_test.go
@@ -12,14 +12,4 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package db
-
-// Emissions define data models
-type Emissions struct {
-	ID            uint    `gorm:"primarykey" json:"id"`
-	ResourceName  string  `gorm:"column:resource_name" json:"resource_name"`
-	ResourceOwner string  `gorm:"column:resource_owner" json:"resource_owner"`
-	ResourceType  string  `gorm:"column:resource_type" json:"resource_type"`
-	Provider      string  `gorm:"column:provider" json:"provider"`
-	MTCO2e        float32 `gorm:"column:mtco2e" json:"mtco2e"`
-}
+package postgres
