+++ pkg/data/db/provider/sqlite/sqlite_test.go
@@ -12,15 +12,4 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package util
-
-import (
-	"testing"
-
-	"github.com/stretchr/testify/assert"
-)
-
-func TestGetDefaultZapCfgPos(t *testing.T) {
-	logger := GetDefaultZapCfg()
-	assert.NotNil(t, logger)
-}
+package sqlite
