+++ pkg/data/db/testdata/not-sqlite.yml
@@ -1,0 +1,2 @@
+file-for-testing: true
+do-not: delete
